+++ test/test-configcatclientintegration.cpp
@@ -1,7 +1,6 @@
 #include <gtest/gtest.h>
 #include "mock.h"
 #include "configcat/configcatclient.h"
-#include "utils.h"
 
 using namespace configcat;
 using namespace std;
