+++ test/test-hooks.cpp
@@ -1,6 +1,5 @@
 #include <gtest/gtest.h>
 #include "mock.h"
-#include "utils.h"
 #include "configservice.h"
 #include "configcat/configcatclient.h"
 #include "configcat/configcatoptions.h"
